modifiche per deploy remoto rev:1.6

+++ backend/copio/src/main/java/com/copio/copio/repository/PageRepository.java
@@ -17,7 +17,7 @@ public interface PageRepository extends JpaRepository <Page,Integer> {
 
     
     // Trova la pagina per idPage e userId
-    Optional<Page> findByIdAndUserId(Long idPage, Long userId);
+    Optional<Page> findByIdAndUserId(Integer idPage, Long userId);
